+++ concrete_datastore/api/v1/filters.py
@@ -334,6 +334,51 @@ class FilterSupportingContainsBackend(
         return queryset.filter(q_object)
 
 
+class FilterSupportingInsensitiveContainsBackend(
+    BaseFilterBackend, CustomShemaOperationParameters
+):
+    def get_schema_operation_parameters(self, view):
+        params = [
+            {
+                'name': f'{field_name}__icontains',
+                'required': False,
+                'in': 'query',
+                'schema': {'type': 'string'},
+            }
+            for field_name in getattr(view, 'filterset_fields', ())
+            if get_filter_field_type(view.model_class, field_name)
+            in ('CharField', 'TextField')
+        ]
+        return self.return_if_not_details(view=view, value=params)
+
+    def filter_queryset(self, request, queryset, view):
+        query_params = request.query_params
+        filterset_fields = getattr(view, 'filterset_fields', ())
+
+        q_object = None
+        for param in query_params:
+            if not param.endswith('__icontains'):
+                continue
+            param_field = param.replace('__icontains', '')
+            if param_field not in filterset_fields:
+                continue
+            if not get_filter_field_type(queryset.model, param_field) in (
+                'CharField',
+                'TextField',
+            ):
+                continue
+
+            custom_filter = {param: query_params.get(param)}
+            if q_object is None:
+                q_object = Q(**custom_filter)
+            else:
+                q_object &= Q(**custom_filter)
+        if q_object is None:
+            return queryset
+
+        return queryset.filter(q_object)
+
+
 class FilterSupportingEmptyBackend(
     BaseFilterBackend, CustomShemaOperationParameters
 ):